statically cache "Editor\getPostModules" results. Every time this function is called it grabs data for all modules and posts, and it's requested for all modules.
As we already get all the data we want the first time,
and the $postId doesn't change in the same instance as far as i know.
We can gain a significant speedboost during loading by doing caching it.
(The more modules used on the post the higher the gain).

remove the unused $available variable.