improve performace when grabbing module posts.
using $wpdb and only getting the ID has quite a bit less overhead compared to "get_posts".

As the results are cached this only affects the first run each instance.

I'm unsure what the preferred formatting is for the project,
and i probably broken some here (at least on the "$format_statement"):p